Adding and editing script, canva
Adding canva for the main menu of the game and the canva to create the
hero. Editing of the script BoardManager to link the interface with the
info about the hero (item stats, hero name, etc). Adding font for the
title of the game. Adding script to make a progress bar for the
progression of the level of our hero, a script to add radio button and a
script to change the source of our hero if it is a wizard or a warrior.